reorganize filesystem
