+++ index.html
@@ -3,7 +3,8 @@
 
 <head>
     <title>TTRPG MINI TOKEN MAKER</title>
-    <link rel="stylesheet" href="upload.css">
+    <link rel="stylesheet" href="src/main.css">
+    <link rel="icon" type="image/png" href="src/favicon.ico">
 </head>
 
 <body>
@@ -13,7 +14,7 @@
     <div id="dropzone" effectAllowed="move">Drop monster images here!</div>
     <ul id="images"></ul>
 
-    <script src="upload.js" type="module"></script>
+    <script src="src/index.js" type="module"></script>
 </body>
 
 </html>
