+++ src/upload.js
@@ -36,16 +36,4 @@ async function readEntriesPromise(directoryReader) {
     }
 }
 
-var elDrop = document.getElementById('dropzone');
-var elItems = document.getElementById('images');
-
-elDrop.addEventListener('dragover', function (event) {
-    event.preventDefault();
-    elItems.innerHTML = 0;
-});
-
-elDrop.addEventListener('drop', async function (event) {
-    event.preventDefault();
-    let items = await getAllFileEntries(event.dataTransfer.items);
-    elItems.innerHTML = items.length;
-});
+export { getAllFileEntries, readAllDirectoryEntries, readEntriesPromise }
